Fix intermittent assertion failure in circular buffer.
On rare occasions, playback was failing just after this assertion failed:

(gst-launch-0.10:2216): GStreamer-CRITICAL **: gst_buffer_create_sub: assertion
`buffer->size >= offset + size' failed

The assertion was failing in gst_ticircbuffer_queue_data, and this is what was
going on.

If the buffer is too big to fit into the remaining space in the circular
buffer, it is split into two sub-buffers -- one that fits exactly in the
remaining space and one that contains the rest of the data.  Then two recursive
calls are made to gst_ticircbuffer_queue_data with these sub-buffers.

This scenario is detected by the following while loop:

    while (gst_ticircbuffer_write_space(circBuf) < GST_BUFFER_SIZE(buf)) {

This detects that the remaining space left is smaller than the buffer size
we're trying to queue.

Inside this loop, we calculate the sizes of the two sub-buffers as follows:

Sub-buffer 1:
   Int32 bytesAvail = gst_ticircbuffer_write_space(circBuf);
   subBuf = gst_buffer_create_sub(buf, 0, bytesAvail);

Sub-buffer 2:
   subBuf = gst_buffer_create_sub(buf, bytesAvail,
                GST_BUFFER_SIZE(buf) - bytesAvail);

Here's the problem:  the first and second calls to gst_ticircbuffer_write_space
may return different values.  If the decode thread consumed data from the
circular buffer when the read pointer is ahead of the write pointer, then
the amount of available space may increase after the initial check in the
while loop.

However, the code _inside_ the while loop assumes that the available space is
always less than the size of the buffer, and if that isn't true, we may end up
trying to create a sub-buffer that is larger than the original buffer.  That's
how we get the assertion failure.

This can be easily remedied.  Once we're committed to filling in the write
space reported in the initial while loop check, we just need to make sure we
use that same value inside the loop.  We simply save off the original value
into a local variable, and use that in the while loop instead of calling
gst_ticircbuffer_write_space again:

   while ((writeSpace = gst_ticircbuffer_write_space(circBuf)) <
           GST_BUFFER_SIZE(buf)) {

Sub-buffer 1:
   subBuf = gst_buffer_create_sub(buf, 0, writeSpace);

Sub-buffer 2:
   subBuf = gst_buffer_create_sub(buf, writeSpace,
                GST_BUFFER_SIZE(buf) - writeSpace);

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@810 24075187-2e39-4e88-bbb8-bc8aa768f540