Replace dFramerateNum and dFramerateDen with dCapsFramerate.
The framerate information we retrieved from the sink caps was being stored
as separate numerator and denominator values in dFramerateNum and
dFramerateDen.  In this change we replace these with a single value
named dCapsFramerate, which is a GstFraction.

This change clarifies the name of the setting so we know where the value
came from, and using a GstFraction is more consistent with other framerate
values after recent changes.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@814 24075187-2e39-4e88-bbb8-bc8aa768f540