Fix handling of framerate values in TIViddec2.
This change allows the framerate property to be specified as a fraction, so
we properly handle fractional framerates like 29.97.

As part of this, the "frameRate" property was renamed to "framerate" to match
the TIDmaiVideoSink, and framerateNum and framerateDen values were replaced
with a GstFraction named "framerate" to move toward using native GStreamer
types.

The "GST_TI_TIViddec2_frameRate" environment variable is still supported, but
there are now additional "GST_TI_TIViddec2_framerateNum" and
"GST_TI_TIViddec2_framerateDen" variables in case you want to specify a
fractional framerate.  These new variables will override
GST_TI_TIViddec2_frameRate if set.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@816 24075187-2e39-4e88-bbb8-bc8aa768f540