Allow size to be specified to gst_tidmaivideosink_alloc_display_buffers
This is a set-up step for supporting pad-allocated buffers.  Upstream elements
may be using codecs that require buffers slightly larger than the default size
for the display.  This change allows the default buffer size to be overridden.
When 0 is specified for the size, the default display buffer size is used.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@820 24075187-2e39-4e88-bbb8-bc8aa768f540