Upgrade GStreamer library to version 0.10.30.
Note that with this GStreamer upgrade, there is now a separate executable
used to scan for available plugins.  For it to work properly, you should
set this variable in your envrionment:

export GST_PLUGIN_SCANNER=${GSTREAMER_DIR}/libexec/gstreamer-0.10/gst-plugin-scanner

GSTREAMER_DIR is location of GStreamer installation on target filesystem, and
by default is "/opt/gstreamer".

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@826 24075187-2e39-4e88-bbb8-bc8aa768f540