dm6467:  Add support for setting v4l2src framerate
The DM6467 capture driver does not let you set the framerate.  Even if you
specify a frame rate of 30/1, you will still get 60/1.  With this change,
if you specify 30/1 and the framerate is really 60/1, every other capture
frame will be dropped and only half of them will be pushed downstream to
simulate 30fps.

The implementation is generic enough to handle any framerate.  It automatically
calculates how many and how often frames need to be dropped.

Unfortuately, there is no way to query the "default" hardware fps for the
capture device, so target-specific code was needed to make this work.
As a result, this feature will only kick-in for DM6467 and DM6467T.  Other
targets could be easily added if needed, but the hardware framerate of the
capture device would need to be added to the code.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@835 24075187-2e39-4e88-bbb8-bc8aa768f540