Remove frame repeat code in TIDmaiVideoSink.
The sync code in the base sink class already delays pushing video frames in
order to match the desired frame rate, as long as the buffers to display have
timestamps.  Because of this, the frame repeat code in TIDmaiVideoSink is
really only useful when there are no timestamps.  I think it's reasonable to
expect that upstream elements set timestamps when throttling the framerate is
important, so this frame repeat code is being removed.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@837 24075187-2e39-4e88-bbb8-bc8aa768f540