Add support for pad-allocated buffers in TIViddec2.
When padAllocOutbufs=TRUE is specified to the TIViddec2 element, it will
attempt to allocate buffers from downstream instead of allocating its own.

Downstream buffers will only be used if it can be determined that they are
all members of a DMAI BufTab, which means they are CMEM-based, and will work
with the codecs.  Currently, the only known downstream element that can provide
these buffers is TIDmaiVideoSink, and it is only supported for DM365.

There is currently no support for centering small clips in the middle of the
display screen -- this would require additional support in the display driver.
As a result, pad-allocation can currently only be used with clips that are at
least as large as the display, and this feature not enabled by default because
of these strict clip-size requirements.

On DM365, there are known issues with the MPEG-2 decoder's output buffer size
calculation that cause it not to work with D1 resolutions unless you hard-code
the size.  H.264 and MPEG-4 decoders work as expected, and MPEG-2 works as
expected for 720p.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@858 24075187-2e39-4e88-bbb8-bc8aa768f540