v4l2src: always set the desired framerate on DM365.
By default, the frame rate is only set if the driver reports that it is
currently using a different value.  The driver doesn't always report the
correct value -- it may report 30fps when it's actually running at 60fps.
This change forces the framerate to always be set, regardless of what the
driver reports.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@859 24075187-2e39-4e88-bbb8-bc8aa768f540