Fix handling of framerate values in TIVidenc1.
This change allows the framerate property to be specified as a fraction, so
we properly handle fractional framerates like 29.97.

As part of this, the "frameRate" property was renamed to "framerate" to match
the TIDmaiVideoSink, and framerateNum and framerateDen values were replaced
with a GstFraction named "framerate" to move toward using native GStreamer
types.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@860 24075187-2e39-4e88-bbb8-bc8aa768f540