Use GstAdapter on TIVidenc1's sink pad.
With a GstAdapter, we can start to make assumptions that the buffers coming
out of the adapter are big enough to feed to the encoder.  In addition,
GstAdapter is also smart enough to give is the direct driver buffer when
v4l2src is using TIVidenc1 for live capture+encode.

Now that the GstAdapter is in place, future changes will take advantage of
the benefits.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@861 24075187-2e39-4e88-bbb8-bc8aa768f540