Cleanup gst_tividenc1_copy_input to use DMAI dst buffer.
We can now assume that gst_tividenc1_copy_input will always give a DMAI buffer
as a destination, so there is no longer a need to create a fake DMAI buffer
when doing a Ccv or Framecopy.  It also helps get the Ccv right on DM6467T, as
the function now gets the metadata directly from the encode function.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@868 24075187-2e39-4e88-bbb8-bc8aa768f540