TIVidenc1: Copy encoded output to a regular GstBuffer
Previously the CMEM-allocated buffers that hold the encoded output were passed
directly downstream.  However, there are only a few of these buffers, so
waiting for one to become available is often more expensive than doing a memcpy
of the encoded frame (which is usually pretty small).

This change copies the encoded data out of the CMEM buffer and passes it
downstream.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@876 24075187-2e39-4e88-bbb8-bc8aa768f540