Correct input buffer size calculation
If we know the input buffer is physically contiguous in memory, use its size as
the expected size for all input buffers.  Otherwise, the codec will be asked
the size to expect.

This helps with encoding 1080p30 on DM6467T, where the codec reports that the
input buffer size should be 1920x1088, where 1920x1080 is usualy what we have
and works fine with the encoder.

Also, on DM365, the input buffers may have addtional padding so the size of
the input buffers is larger than what the codec needs.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@879 24075187-2e39-4e88-bbb8-bc8aa768f540