Enable pad-allocated buffers for OMAP35x/DM37x.
Requires preliminary support for DVSDK 4.00 (omap3530_dv400 build target).

Rotation is not allowed when pad allocation is in use.  When rotation is used,
we can only allocate a maximum of four V4L2 buffer handles.  This isn't
practical when sharing buffers with codecs that could require more buffers than
this.

git-svn-id: https://gstreamer.ti.com/svn/gstreamer_ti/trunk@895 24075187-2e39-4e88-bbb8-bc8aa768f540